Refactor and restructure.
