+++ client/listeners/resetListeners.js
@@ -1,22 +1,9 @@
 const { addLoadRef } = require('./../state/elementRefs')
-const state = require('./../state/state.js')
-const addTrackForm = require('./../utils/addTrackForm.js')
 
-function resetTracklist() {
-  const $tracklistError = addLoadRef('tracklist-error-message-container')
-  const $trackFormContainer = addLoadRef('track-form-container')
-
-  $tracklistError.textContent = ''
-  $trackFormContainer.innerHTML = ''
-  state.currentTrack = 1
-  addTrackForm(state.currentTrack)
-  state.currentTrack += 1
-}
+const resetTracklist = require('./../utils/resetTracklist')
 
 function startOver() {
-  const $trackFinalContainer = addLoadRef('track-final-container')
-
-  $trackFinalContainer.innerHTML = ''
+  resetTracklist()
   window.location.hash = ''
 }
 
@@ -26,7 +13,6 @@ function createResetListeners() {
 
   const attachStartOverBtnListener = () => (
     $startOverBtn.addEventListener('click', () => {
-      resetTracklist()
       startOver()
     })
   )
